requirements.txt
streamlit
solana
pillow
fastapi
uvicorn
pydantic